Moved annotations of spring-boot-vaadin => org.vaadin.spring.boot.annotation
